# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Badminton Court Booking API",
        "description": "API untuk sistem pemesanan lapangan badminton",
        "contact": {
            "name": "Badminton Booking Support",
            "email": "admin@badminton-booking.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Development Server"
        }
    ],
    "paths": {
...
        "/api/v1/schedules": {
            "get": {
                "tags": [
                    "Schedules"
                ],
                "summary": "Dapatkan semua jadwal",
                "description": "Mengambil semua jadwal lapangan badminton yang diurutkan berdasarkan tanggal dan waktu mulai",
                "operationId": "209f009c324e040a1cf5a0b9f4b6b87c",
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data jadwal",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Schedules retrieved successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Schedule"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Schedules"
                ],
                "summary": "Buat jadwal baru",
                "description": "Membuat jadwal baru untuk lapangan badminton",
                "operationId": "f28e57e7415c1768bc469b4bf4d3fdaf",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "court_id",
                                    "date",
                                    "start_time",
                                    "end_time"
                                ],
                                "properties": {
                                    "court_id": {
                                        "description": "ID lapangan",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "date": {
                                        "description": "Tanggal jadwal",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-01-15"
                                    },
                                    "start_time": {
                                        "description": "Waktu mulai (format HH:mm)",
                                        "type": "string",
                                        "format": "time",
                                        "example": "09:00"
                                    },
                                    "end_time": {
                                        "description": "Waktu selesai (format HH:mm)",
                                        "type": "string",
                                        "format": "time",
                                        "example": "11:00"
                                    },
                                    "status": {
                                        "description": "Status jadwal",
                                        "type": "string",
                                        "enum": [
                                            "available",
                                            "booked"
                                        ],
                                        "example": "available"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Jadwal berhasil dibuat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Schedule created successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Schedule"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/schedules-available": {
            "get": {
                "tags": [
                    "Schedules"
                ],
                "summary": "Dapatkan jadwal yang tersedia",
                "description": "Mengambil semua jadwal lapangan yang memiliki status 'available' untuk booking",
                "operationId": "a1274e2b34d8d4e5081acdc79e1701cf",
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil jadwal yang tersedia",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Available schedules retrieved successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Schedule"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/schedules/{id}": {
            "put": {
                "tags": [
                    "Schedules"
                ],
                "summary": "Update jadwal",
                "description": "Mengupdate data jadwal lapangan badminton",
                "operationId": "1f8946f70306e18b2405f59d67f94a13",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID jadwal",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "court_id": {
                                        "description": "ID lapangan",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "date": {
                                        "description": "Tanggal jadwal",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-01-15"
                                    },
                                    "start_time": {
                                        "description": "Waktu mulai (format HH:mm)",
                                        "type": "string",
                                        "format": "time",
                                        "example": "09:00"
                                    },
                                    "end_time": {
                                        "description": "Waktu selesai (format HH:mm)",
                                        "type": "string",
                                        "format": "time",
                                        "example": "11:00"
                                    },
                                    "status": {
                                        "description": "Status jadwal",
                                        "type": "string",
                                        "enum": [
                                            "available",
                                            "booked"
                                        ],
                                        "example": "available"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Jadwal berhasil diupdate",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Schedule updated successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Schedule"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Jadwal tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Schedules"
                ],
                "summary": "Hapus jadwal",
                "description": "Menghapus jadwal lapangan badminton",
                "operationId": "276c2e51a02028c1ff6417817dc3a0e7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID jadwal",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Jadwal berhasil dihapus",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Schedule deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Jadwal tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Court": {
                "title": "Court",
                "description": "Model lapangan badminton",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Lapangan A"
                    },
                    "description": {
                        "type": "string",
                        "example": "Lapangan premium dengan lantai kayu"
                    },
                    "price_per_hour": {
                        "type": "number",
                        "format": "decimal",
                        "example": 50000
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "active",
                            "inactive"
                        ],
                        "example": "active"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-15T10:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-15T10:00:00Z"
                    }
                },
                "type": "object"
            },
            "Schedule": {
                "title": "Schedule",
                "description": "Model jadwal lapangan",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "court_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "date": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-01-15"
                    },
                    "start_time": {
                        "type": "string",
                        "format": "time",
                        "example": "09:00"
                    },
                    "end_time": {
                        "type": "string",
                        "format": "time",
                        "example": "11:00"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "available",
                            "booked"
                        ],
                        "example": "available"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-15T10:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-15T10:00:00Z"
                    },
                    "court": {
                        "$ref": "#/components/schemas/Court"
                    }
                },
                "type": "object"
            },
...
            "ErrorResponse": {
                "title": "Error Response",
                "description": "Error response format",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Error occurred"
                    },
                    "error": {
                        "type": "string",
                        "example": "Detailed error message"
                    }
                },
                "type": "object"
            },
            "ValidationErrorResponse": {
                "title": "Validation Error Response",
                "description": "Validation error response format",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "properties": {
                            "field_name": {
                                "type": "array",
                                "items": {
                                    "type": "string",
                                    "example": "The field is required."
                                }
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            }
        }
    },
...
}